Remove `Exports.WindowState.cpp` and related `LibraryImports` for unused window state interop APIs in `NativeBridge`.
