+++ src/InfiniFrame.NativeBridge/Managed/NativeApi/InfiniFrameNativeInteropStatus.cs
