+++ src/InfiniFrame.NativeBridge/Native/src/Api/Exports/Exports.Platform.MacOs.cpp
@@ -1,0 +1,14 @@
+// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+#include "Api/Exports/Exports.h"
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+extern "C" {
+#ifdef __APPLE__
+EXPORTED InteropStatus InfiniFrame_register_mac() {
+    return RunExportStatus([] { InfiniFrameWindow::Register(); });
+}
+#endif
+}
